ENH: More work on caching.
This time removed prefetching of the cache. Before I was prefetching
the cache, in order to decrease load time of objects in the future.
But this was not good or correct. Now the template body is only downloaded
when needed. This improves startup time, by traiding some safty. Anyway,
this should speed up the load time.